requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
peft>=0.7.0

# ControlNet and image processing
controlnet-aux>=0.0.9
pillow>=10.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Training utilities
datasets>=2.14.0
xformers>=0.0.22
safetensors>=0.4.0

# Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For API (if needed)
# Uncomment to enable API server
# fastapi>=0.104.0
# uvicorn>=0.24.0
# python-multipart>=0.0.6
# pydantic>=2.0.0

